+++ Assets/Scripts/Blindsided/UGS/UgsLeaderboardIds.cs
@@ -6,10 +6,9 @@ namespace Blindsided.UGS
     /// </summary>
     public static class UgsLeaderboardIds
     {
-        public const string SpookyKills = "SpookyKills";
-        public const string SpookyKillsSeasonal = "Spooky_Kills_Seasonal";
-        public const string DistanceTravelled = "Distance_Travelled";
-        public const string Kills = "Kills";
+        public const string CompletionTime = "Completion_TIme";
+        public const string CompletionTimeCheaters = "Completion_Time_Cheaters";
+        public const string DistanceReached = "Distance_Reached";
         public const string Tasks = "Tasks";
     }
 }